Disable terminal functionality
I removed most of the terminal-related contributions (terminal.contribution.ts), which includes: commands, the terminal tab, menu items, keybindings, and so on.

Many terminal-related services are used by other services (e.g. the task service depends on the terminal service), so removing TerminalService would cause a cascade of "missing singleton" errors.

Thankfully vscode uses DI everywhere,  so I replaced the TerminalService with our own empty shell of a service. Same for a few related services. The empty shells are there to satisfy the dependency injection system.

Note that there's still quite a bit of terminal-related code left that I think will end up in our build. We might be able to use a webpack plugin to for example, replace xterm.js with a dummy module.